feat: add option to set a custom lipgloss renderer
Lip Gloss defaults to `os.Stdout` to detect the background color and
color profile of the terminal. This can be a problem when the
application is using a different output stream like `os.Stderr` or a
remote session like in a SSH server.

When using an SSH server, usually the server provides to the session a
`$SSH_TTY` environment variable pointing to the allocated PTY to be used
for the application. The value of `$SSH_TTY` points to the terminal TTY
  FD that _should_ be used to run the application.

Starting with Lip Gloss v0.7, you can use custom renderers to specify a
different Lip Gloss output to detect the background color and color
profile. This PR adds the necessary "glue" to the Bubbles in order to
change the renderer being used to detect the color profile and
background.