fix: handle textinput max-width edge cases
I noticed that the cursor seemed to be adding an extra space which the
width calculations weren't accounting for. This also led me to discover
an edge case with the view (`offset`/`offsetRight`) calculation when the
cursor (`pos`) sits at the very end. Additionally, `View()` was always
rendering the cursor, even when `focus` was false, which the comment
suggests should not be the case.

These should all be resolved, such that Width is now fully respected in
both `View()` and `placeholderView()`: they should _never_ render a
string that exceeds the requested width. (Well... with one caveat: I did
not touch the suggestion logic, as I've never used it. I did leave a
comment, however, that perhaps suggestions shouldn't even render when
the component does not have focus.)

Fixes #358, #307