Improve textarea tests
Improve textarea tests to check the cursor location. This is the
displayed location and not the line row and column (which does not take
into account soft wrapping).

As part of this improvement, a number of existing tests were merged
into the view test tables as it was unnecessary duplication of code
having them separated.

Signed-off-by: Michael Lorant <michael.lorant@nine.com.au>