Fix end of buffer character
The end of buffer character was not being rendered when line numbers
were not enabled.

When the new test tables unit tests were created this issue was
identified however to limit the scope of those changes to just creating
unit tests, this regression was maintained.

This change fixes this issue and updates the tests to the intended
result.

Signed-off-by: Michael Lorant <michael.lorant@nine.com.au>