feat(textinput): expose `UpdateSuggestions` method
This enables clients of the lib to update the suggestions at runtime
whenever they please.

One direct application of this is when using `m.SetValue()`:
After calling it, there is no way to get the updated
`m.CurrentSuggestion()` as `m.SetValue()` does not update the suggestion
list.

It would be possible to also run `m.updateSuggestions()` somewhere in
`m.SetValue`, but that might interfere with other things.