feat(textinput): store and recover history
This adds the history functionality to the textinput model. Specific
actions can push a new value into the history and they can be switched
back to via key presses.

This is done in relation to https://github.com/nobe4/gh-not/issues/190
I was considering implementing that in my project, and realized that it
could just be part of bubbles instead.

Looking forward to getting some feedback on this. Especially the
comments below.