feat: use inhouse wcwidth package instead of go-runewidth
This change replaces the `go-runewidth` package with our in-house
x/wcwidth package. This package is a simplified version of go-runewidth
that uses the `golang.org/x/text` package for Unicode character width
calculations.

We also can't be using both go-runewidth and uniseg to calculate
character widths, so we've removed the uniseg package as well.
Otherwise, we might end up with inconsistent character width
in terminals that don't support grapheme clusters.