feat(filepicker): add support for abstracted filesystem io/fs.FS
Add an optional io/fs.FS to filepicker's model to allow to browse an
abstracted filesystem instead of the full filesystem exposed by the 'os'
package.

Note that Go 1.25 will add io/fs.ReadLink, but we can't use it yet. So
we implement symlink resolution here.

By the way, references to os types and constants which are aliases to
same symbols in io/fs are replaced: os.Mode*, os.DirEntry.