fix(textarea): ensure viewport content is set during update
This ensures that the viewport content is set and persisted correctly.
Setting the content during View won't doesn't work because it mutates a
copy of the model instead without persisting the changes. Moreover, any
after effects of setting the content won't be visible because the next
Update will be called on the older model, not the new one.